pyproject.toml
[project]
name = "minimax-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiodns>=3.5.0",
    "aiohttp>=3.12.14",
    "fastapi[standard]>=0.116.1",
    "nanoid>=2.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
]
